+++ src/detectors/index.ts
@@ -14,6 +14,7 @@ import { handlerDepthDetector } from "./handlerDepth.js";
 import { hookDependencySmellDetector } from "./hookDependencySmell.js";
 import { importCycleDetector } from "./importCycle.js";
 import { kotlinDeadAbstractionDetector, kotlinDuplicateLogicDetector, kotlinLargeFunctionDetector, kotlinTodoCommentDetector } from "./kotlin/index.js";
+import { swiftDeadAbstractionDetector, swiftDuplicateLogicDetector, swiftLargeFunctionDetector, swiftTodoCommentDetector } from "./swift/index.js";
 import { largeComponentDetector } from "./largeComponent.js";
 import { largeFunctionDetector } from "./largeFunction.js";
 import { namingDriftDetector } from "./namingDrift.js";
@@ -81,6 +82,10 @@ export const allDetectors: Detector[] = [
   kotlinLargeFunctionDetector,
   kotlinDuplicateLogicDetector,
   kotlinDeadAbstractionDetector,
+  swiftTodoCommentDetector,
+  swiftLargeFunctionDetector,
+  swiftDuplicateLogicDetector,
+  swiftDeadAbstractionDetector,
   composeLargeComposableDetector,
   composeStateHoistingDetector,
 ];